Rename `molecules_old` to `molecules_jcesr` (#423)
* Rename molecules_old to molecules_jcesr

* Propagate molecules rename to api
