+++ emmet-api/emmet/api/routes/molecules/resources.py
@@ -1,5 +1,5 @@
 from maggma.api.resource import ReadOnlyResource
-from emmet.core.molecules_old import MoleculesDoc
+from emmet.core.molecules_jcesr import MoleculesDoc
 
 from maggma.api.query_operator import PaginationQuery, SortQuery, SparseFieldsQuery
 from emmet.api.routes.molecules.query_operators import (
